+++ bridgetown-core/test/dest/dynamic_file.php
@@ -1,1 +1,0 @@
-I'm a Bridgetown file! I should be output as dynamic_file.php, no .html to be found.