+++ bridgetown-website/frontend/javascript/index.js
@@ -1,0 +1,3 @@
+import "../styles/index.scss"
+
+console.info("Bridgetown is loaded!")