+++ History.markdown
@@ -1,5 +1,10 @@
 ## HEAD
 
+## 0.4.0 / 2020-04-05
+
+  * Added a `component` Liquid tag which extends the functionality of include tags.
+  * Added a new `bridgetown-website` project to the repo, which of course is a Bridgetown site and will house the homepage, documentation, etc.
+
 ## 0.3.0 / 2020-04-05
 
   * Moved all Bridgetown code to `bridgetown-core`, the idea being this will now be a monorpo housing Core plus a few other official gems/projects as time goes on. Users will install the `bridgetown` gem which in turns installs `bridgetown-core` as a dependency.